Rename: dto를 request response로 구분하게 변경
