+++ src/main/java/beachcombine/backend/dto/request/MemberUpdateRequest.java
@@ -1,4 +1,4 @@
-package beachcombine.backend.dto.member;
+package beachcombine.backend.dto.request;
 
 import lombok.*;
