+++ src/main/java/beachcombine/backend/service/MemberService.java
@@ -3,8 +3,8 @@ package beachcombine.backend.service;
 import beachcombine.backend.common.exception.CustomException;
 import beachcombine.backend.common.exception.ErrorCode;
 import beachcombine.backend.domain.Member;
-import beachcombine.backend.dto.member.MemberResponse;
-import beachcombine.backend.dto.member.MemberUpdateRequest;
+import beachcombine.backend.dto.response.MemberResponse;
+import beachcombine.backend.dto.request.MemberUpdateRequest;
 import beachcombine.backend.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
